Exported GachaItem to gacha-details
